+++ lib/rom/sql/function.rb
@@ -26,6 +26,21 @@ module ROM
         ::Sequel::SQL::BooleanExpression.new(:'=', func, other)
       end
 
+      # Convert an expression result to another data type
+      #
+      # @example
+      #   users.select { bool::cast(json_data.get_text('activated'), :boolean).as(:activated) }
+      #
+      # @param [ROM::SQL::Attrbite] expr Expression to be cast
+      # @param [String] db_type Target database type
+      #
+      # @return [ROM::SQL::Attribute]
+      #
+      # @api private
+      def cast(expr, db_type)
+        Attribute[type].meta(sql_expr: ::Sequel.cast(expr, db_type))
+      end
+
       private
 
       def func